+++ CHANGELOG.md
@@ -4,12 +4,16 @@
 
 ### Features
 
+- [#1413](https://github.com/influxdata/kapacitor/issues/1413): Add subscriptions modes to InfluxDB subscriptions.
+
 ### Bugfixes
 
 ## v1.3.2 [2017-08-08]
 
 ### Bugfixes
 - [#1512](https://github.com/influxdata/kapacitor/pull/1512): Use details field from alert node in PagerDuty.
+- [#1400](https://github.com/influxdata/kapacitor/issues/1400): Allow for `.yml` file extensions in `define-topic-handler`
+- [#1402](https://github.com/influxdata/kapacitor/pull/1402): Fix http server error logging.
 
 ## v1.3.1 [2017-06-02]